requirements.txt
python-dotenv
xai-sdk
requests 
python-amazon-paapi # amazon_paapi
